Add files via upload
This is a Automatic file Sorting Project. 

I have written a single python script which will help compartmentalize all the different files in the specific folders you want to store in. It is helpful when you have multiple file that you wanna move to different folders from time to time. You can set up the time to refresh and run this script and every thing will happen automatically in the background. 

You are Welcome!